requirements.txt
django==5.1.3
pipdeptree==2.23.4
mysqlclient==2.2.5
djangorestframework==3.15.2
